error handling during image deletion

+++ src/controllers/adminControllers/bannerControllers.js
@@ -118,17 +118,23 @@ export const editBanner = async (req, res, next) => {
 export const deleteBannerImage = async (req, res, next) => {
     const { id } = req.params;
     const { image } = req.body;
-    
+
     try {
         await Banner.findByIdAndUpdate(id, { $pull: { images: image } }, { new: true });
 
-        fs.unlink(path.join(__dirname, "../../public/uploads", image), (err) => {
-            if (err) {
-                throw err;
-            } else {
-                res.redirect(`/admin/banners/${id}/edit`);
-            }
-        });
+        // Construct the full path to the image file
+        const imagePath = path.join(__dirname, "../../public/uploads", image);
+
+        // Check if the image file exists
+        if (fs.existsSync(imagePath)) {
+            fs.unlink(path.join(__dirname, "../../public/uploads", image), (err) => {
+                if (err) {
+                    throw err;
+                }
+            });
+        }
+
+        res.redirect(`/admin/banners/${id}/edit`);
     } catch (error) {
         next(error);
     }
